update file and optimize dashboard quickly view
