+++ src/app/layout.js
@@ -1,20 +1,10 @@
-'use client';
-import Footer from "@/components/global/Footer";
-import { ThemeProvider } from "@/components/theme-provider";
-import { LoginUserToDB } from "@/lib/loginUserToDB";
-import { OfflineUserToDB } from "@/lib/offlineUserToDB";
-import { SaveUserToDB } from "@/lib/saveUserToDB";
+"use client";
+
 import Providers from "@/providers/Providers";
-import { ClerkProvider } from "@clerk/nextjs";
-import { Geist, Geist_Mono } from "next/font/google";
-import { usePathname } from "next/navigation";
-import { useEffect, useState } from "react";
-import { Toaster } from "sonner";
-import "./globals.css";
-import { XynexaNavbar } from "@/components/global/XynexaNavbar";
 import { HMSRoomProvider } from "@100mslive/react-sdk";
-//   description: "Generated by Debug Titans Team",
-// };
+import { ClerkProvider } from "@clerk/nextjs";
+import ContentWithLogic from "./ContentWithLogic";
+import { Geist, Geist_Mono } from "next/font/google";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
@@ -27,60 +17,25 @@ const geistMono = Geist_Mono({
 });
 
 export default function RootLayout({ children }) {
-  const pathname = usePathname();
-  const [isDashboard, setIsDashboard] = useState(false);
-  const [isAdminDashboard, setIsAdminDashboard] = useState(false);
-  const [isSignIn, setIsSignIn] = useState(false);
-  const [isSignUp, setIsSignUp] = useState(false);
-  const [shouldShowNavbarFooter, setShouldShowNavbarFooter] = useState(true);
-
-  useEffect(() => {
-    setIsDashboard(pathname.startsWith("/dashboard"));
-    setIsAdminDashboard(pathname.startsWith("/admin-dashbaord")); // Corrected path
-    setIsSignIn(pathname.startsWith("/sign-in"));
-    setIsSignUp(pathname.startsWith("/sign-up"));
-  }, [pathname]);
-
-  useEffect(() => {
-    setShouldShowNavbarFooter(!(isDashboard || isAdminDashboard || isSignIn || isSignUp));
-  }, [isDashboard, isAdminDashboard, isSignIn, isSignUp]);
-
   return (
     <HMSRoomProvider>
-        <ClerkProvider
-          appearance={{
-            layout: {
-              unsafe_disableDevelopmentModeWarnings: true,
-            },
-          }}
-        >
-          <html lang="en">
-            <body
-              className={`${geistSans.variable} ${geistMono.variable} antialiased`}
-            >
-              <Providers>
-                <ThemeProvider
-                  attribute="class"
-                  defaultTheme="system"
-                  enableSystem
-                  disableTransitionOnChange
-                >
-                  {shouldShowNavbarFooter && <XynexaNavbar />}
-                  <main className="min-h-screen">
-                    <SaveUserToDB />
-                    <OfflineUserToDB />
-                    <LoginUserToDB />
-
-                    <Toaster position="top-right" />
-
-                    {children}
-                  </main>
-                  {shouldShowNavbarFooter && <Footer />}
-                </ThemeProvider>
-              </Providers>
-            </body>
-          </html>
-        </ClerkProvider>
+      <ClerkProvider
+        appearance={{
+          layout: {
+            unsafe_disableDevelopmentModeWarnings: true,
+          },
+        }}
+      >
+        <html lang="en">
+          <body
+            className={`${geistSans.variable} ${geistMono.variable} antialiased`}
+          >
+            <Providers>
+              <ContentWithLogic>{children}</ContentWithLogic>
+            </Providers>
+          </body>
+        </html>
+      </ClerkProvider>
     </HMSRoomProvider>
   );
 }
